+++ src/includes/functions.ts
@@ -99,3 +99,8 @@ export function string2filename(str: string): string
 
 	return str.trim();
 }
+
+export function delay(ms: number): Promise<void>
+{
+	return new Promise(resolve => setTimeout(resolve, ms));
+}