+++ changelog.md
@@ -3,9 +3,12 @@
 ## 1.0.8
 
 - action: default icon changed to the off one
+- actions: added progressbar for the long operation
 - options: added warning for the "Suspend file:// tabs" option
+- options: added progressbar for the Migrate process
 - readme: added the "Suspend file:// tabs" option
- 
+- suspended: added "Set keyboard shortcuts" link
+
 ## 1.0.7
 
 - bugfix: file:// permission cannot be requested